+++ src/spycipdb/clis/cli_noe.py
@@ -1,5 +1,5 @@
 """
-Back-calculates the PRE distances from PDB structure file.
+Back-calculates the NOE distances from PDB structure file.
 
 Uses idpconfgen libraries for coordinate parsing as it's proven
 to be faster than BioPython.
@@ -8,13 +8,13 @@ Back-calculator logic inspired from X-EISD.
 Error = 0.0001 as reported in Lincoff et al. 2020.
 
 USAGE:
-    $ spycipdb pre <PDB-FILES> [--exp-file]
-    $ spycipdb pre <PDB-FILES> [--exp-file] [--output] [--ncores]
+    $ spycipdb noe <PDB-FILES> [--exp-file]
+    $ spycipdb noe <PDB-FILES> [--exp-file] [--output] [--ncores]
 
 REQUIREMENTS:
     Experimental data must be comma-delimited with at least the following columns:
     
-    res1,atom1,res2,atom2
+    res1,atom1,atom1_multiple_assignments,res2,atom2,atom2_multiple_assignments
     
     Where res1/atom1 is the atom number and name respectively for the first residue
     and res2/atom2 is the atom number and name respectively for the second residue.
@@ -25,7 +25,13 @@ OUTPUT:
     atom-names:
     
     {
-        'format': {'res1': [], 'atom1': [], 'res2': [], 'atom2': []},
+        'format': { 'res1': [],
+                    'atom1': [],
+                    'atom1_multiple_assignments': [],
+                    'res2': [],
+                    'atom2': [],
+                    'atom2_multiple_assignments': []
+                    },
         'pdb1': [dist_values],
         'pdb2': [dist_values],
         ...
@@ -41,7 +47,7 @@ from functools import partial
 from spycipdb import log
 from spycipdb.libs import libcli
 from spycipdb.logger import S, T, init_files, report_on_crash
-from spycipdb.libs.libfuncs import get_scalar, get_pdb_paths
+from spycipdb.libs.libfuncs import get_pdb_paths, get_scalar
 
 from idpconfgen.libs.libmulticore import pool_function
 from idpconfgen.libs.libstructure import(
@@ -50,9 +56,9 @@ from idpconfgen.libs.libstructure import(
     col_resSeq,
     )
 
-LOGFILESNAME = '.spycipdb_pre'
-_name = 'pre'
-_help = 'PRE back-calculator given experimental data template.'
+LOGFILESNAME = '.spycipdb_noe'
+_name = 'noe'
+_help = 'NOE back-calculator given experimental data template.'
 
 _prog, _des, _usage = libcli.parse_doc_params(__doc__)
 
@@ -86,16 +92,19 @@ def get_exp_format_pre(fexp):
     
     format['res1'] = exp.res1.values.astype(int).tolist()
     format['atom1'] = exp.atom1.values.tolist()
+    format['atom1_multiple_assignments'] = exp.atom1_multiple_assignments.values.tolist()
     format['res2'] = exp.res2.values.astype(int).tolist()
     format['atom2'] = exp.atom2.values.tolist()
+    format['atom2_multiple_assignments'] = exp.atom2_multiple_assignments.values.tolist()
     
     return format
 
 
-def calc_pre(fexp, pdb):
+def calc_noe(fexp, pdb):
     """
-    Main logic for back-calculating PRE data
-    with atom-pairs derived from experimental template.
+    Main logic for back-calculating NOE data
+    with atom-pairs and multi-assigns derived
+    from experimental template
     """
     dist = []
     
@@ -104,33 +113,51 @@ def calc_pre(fexp, pdb):
     atom1_name = exp.atom1.values
     res2 = exp.res2.values.astype(int)
     atom2_name = exp.atom2.values
+    multi1 = exp.atom1_multiple_assignments.values
+    multi2 = exp.atom2_multiple_assignments.values
     
     s = Structure(pdb)
     s.build()
     
-    for i in range (exp.shape[0]):
+    for i in range(exp.shape[0]):
         r1 = int(res1[i])
         r2 = int(res2[i])
+        atom1_list = []
+        atom2_list = []
         for j, r in enumerate(s.data_array[:, col_resSeq].astype(int)):
             if r == r1:
                 if atom1_name[i] == 'H':
-                    atom1 = s.coords[j, :]
+                    atom1_list.append(s.coords[j, :])
                     break
                 if atom1_name[i] in s.data_array[j, col_name]:
-                    atom1 = s.coords[j, :]
+                    atom1_list.append(s.coords[j, :])
+                if len(atom1_list) == 2:
+                    break
+                if not multi1[i] and len(atom1_list) == 1:
                     break
         for j, r in enumerate(s.data_array[:, col_resSeq].astype(int)):
             if r == r2:
                 if atom2_name[i] == 'H':
-                    atom2 = s.coords[j, :]
+                    atom2_list.append(s.coords[j, :])
                     break
                 if atom2_name[i] in s.data_array[j, col_name]:
-                    atom2 = s.coords[j, :]
+                    atom2_list.append(s.coords[j, :])
+                if len(atom2_list) == 2:
                     break
-                
-        dv = atom1 - atom2
-        assert dv.shape == (3,)
-        dist.append(get_scalar(dv[0], dv[1], dv[2]))
+                if not multi2[i] and len(atom2_list) == 1:
+                    break
+
+        combos = 0.0
+        num_combos = 0
+
+        for first_atom in atom1_list:
+            for second_atom in atom2_list:
+                dv = first_atom - second_atom
+                assert dv.shape == (3,)
+                combos += (get_scalar(dv[0], dv[1], dv[2])) ** (-6.)
+                num_combos += 1
+
+        dist.append((combos / float(num_combos)) ** (-1 / 6))
     
     return pdb, dist
 
@@ -144,9 +171,9 @@ def main(
         **kwargs,
         ):
     """
-    Main logic for back-calculating PRE values from PDB structures
-    given experimental file template.
-
+    Main logic for processing PDB structures and
+    outputting back-calculated NOE values.
+    
     Parameters
     ----------
     pdb_files : str or Path, required
@@ -154,7 +181,7 @@ def main(
         
     exp_file : str or Path, required
         Path to experimental file template.
-        Required to know which distances to calculate.
+        Required to know for which residues to calculate.
     
     output : str or Path, optional
         Where to store the back-calculated data.
@@ -169,7 +196,6 @@ def main(
         Defaults to TMPDIR.
     """
     init_files(log, LOGFILESNAME)
-    
     log.info(T('reading input paths'))
     pdbs2operate, _istarfile = get_pdb_paths(pdb_files, tmpdir)
     log.info(S('done'))
@@ -177,7 +203,7 @@ def main(
     log.info(T(f'back calculating using {ncores} workers'))
     execute = partial(
         report_on_crash,
-        calc_pre,
+        calc_noe,
         exp_file,
         )
     execute_pool = pool_function(execute, pdbs2operate, ncores=ncores)
@@ -193,7 +219,7 @@ def main(
         fout.write(json.dumps(_output, indent=4))
     log.info(S('done'))
     
-
+    
     if _istarfile:
         shutil.rmtree(tmpdir)